@@ -1,12 +1,14 @@
 import 'dart:async';
-import 'dart:collection';
 import 'dart:developer' show log;
 
-import 'package:flutter/foundation.dart';
-import 'package:http/http.dart' as http;
+import 'package:collection/collection.dart';
 import 'package:csv/csv.dart';
+import 'package:dart_rss/dart_rss.dart';
+import 'package:flutter/foundation.dart';
+import 'package:flutter/services.dart' show rootBundle;
+import 'package:http/http.dart' as http;
 import 'package:intl/intl.dart';
-import 'package:latlong/latlong.dart';
+import 'package:latlong2/latlong.dart';
 import 'package:snow_weather_info/data/sources/avalanche_api.dart';
 import 'package:snow_weather_info/data/sources/data_api.dart';
 import 'package:snow_weather_info/data/sources/database_helper.dart';
@@ -14,17 +16,17 @@ import 'package:snow_weather_info/data/sources/preferences.dart';
 import 'package:snow_weather_info/model/avalanche_bulletin.dart';
 import 'package:snow_weather_info/model/data_station.dart';
 import 'package:snow_weather_info/model/station.dart';
-import 'package:dart_rss/dart_rss.dart';
+
 import 'constant_data_list.dart';
 
 class DataNotifier extends ChangeNotifier {
-  Preferences preferences;
-  AvalancheAPI avalancheAPI;
-  DataAPI dataAPI;
-  DatabaseHelper databaseHelper;
+  late Preferences preferences;
+  late AvalancheAPI avalancheAPI;
+  late DataAPI dataAPI;
+  late DatabaseHelper databaseHelper;
 
   bool _isInitialise = false;
-  HashMap<int, List<DataStation>> _mapDataStation;
+  final _mapDataStation = <int, List<DataStation>>{};
   LatLng currentMapLoc = LatLng(45.05, 6.3);
   int currentIndexTab = 0;
 
@@ -42,7 +44,7 @@ class DataNotifier extends ChangeNotifier {
   }
 
   List<Station> get stations => _stations.toList();
-  List<Station> _stations;
+  List<Station> _stations = [];
   @protected
   set stations(List<Station> value) {
     if (_stations != value) {
@@ -52,7 +54,7 @@ class DataNotifier extends ChangeNotifier {
   }
 
   List<Nivose> get nivoses => _nivoses.toList();
-  List<Nivose> _nivoses;
+  List<Nivose> _nivoses = [];
   @protected
   set nivoses(List<Nivose> value) {
     if (_nivoses != value) {
@@ -64,7 +66,7 @@ class DataNotifier extends ChangeNotifier {
   List<AbstractStation> get allStations => [..._nivoses, ..._stations];
 
   List<AbstractStation> get favoritesStations => _favoritesStations.toList();
-  List<AbstractStation> _favoritesStations;
+  List<AbstractStation> _favoritesStations = [];
   @protected
   set favoritesStations(List<AbstractStation> value) {
     if (_favoritesStations != value) {
@@ -73,10 +75,10 @@ class DataNotifier extends ChangeNotifier {
     }
   }
 
-  AtomFeed get avalancheFeed => _avalancheFeed;
-  AtomFeed _avalancheFeed;
+  AtomFeed? get avalancheFeed => _avalancheFeed;
+  AtomFeed? _avalancheFeed;
   @protected
-  set avalancheFeed(AtomFeed value) {
+  set avalancheFeed(AtomFeed? value) {
     if (_avalancheFeed != value) {
       _avalancheFeed = value;
       notifyListeners();
@@ -84,7 +86,7 @@ class DataNotifier extends ChangeNotifier {
   }
 
   List<DataStation> getDataOfStation(int id) {
-    return _mapDataStation[id];
+    return _mapDataStation[id] ?? [];
   }
 
   Nivose getNivose(String codeMF) {
@@ -101,6 +103,8 @@ class DataNotifier extends ChangeNotifier {
         await Future.wait([
           _initStation(),
           _downloadStationData(),
+          // uncomment to use test value
+          //_readTestStationData(),
         ]);
         avalancheFeed = await avalancheAPI.getAvalanche();
         await _finalizeStationData();
@@ -117,7 +121,6 @@ class DataNotifier extends ChangeNotifier {
   }
 
   Future<void> _finalizeStationData() async {
-    _mapDataStation = HashMap();
     await databaseHelper.cleanOldData(7);
 
     for (final s in _stations) {
@@ -125,10 +128,8 @@ class DataNotifier extends ChangeNotifier {
       _mapDataStation[s.id] = listOfData;
       s.hasData = listOfData.isNotEmpty;
       if (s.hasData) {
-        final dataSt =
-            listOfData.firstWhere((d) => d.hasSnowHeight, orElse: () => null);
-        s.lastSnowHeight =
-            dataSt != null && dataSt.hasSnowHeight ? dataSt.snowHeight : 0.0;
+        final dataSt = listOfData.firstWhereOrNull((d) => d.snowHeight != null);
+        s.lastSnowHeight = dataSt?.snowHeight ?? 0;
       } else {
         s.lastSnowHeight = 0.0;
       }
@@ -181,8 +182,32 @@ class DataNotifier extends ChangeNotifier {
     );
   }
 
+  void _decodeStationData(String data) {
+    final cvsResult = const CsvToListConverter().convert(
+      data,
+      fieldDelimiter: ';',
+      eol: '\n',
+      shouldParseNumbers: false,
+    );
+
+    if (cvsResult.length > 1) {
+      if (cvsResult[0].length > 2) {
+        cvsResult.removeAt(0);
+        final listData = cvsResult.map<DataStation>(
+          (line) {
+            return DataStation.fromList(line);
+          },
+        ).toList();
+
+        listData.forEach((d) async {
+          await databaseHelper.insertStationData(d);
+        });
+      }
+    }
+  }
+
   Future<void> _downloadStationData() async {
-    DateTime lastDataDownload;
+    DateTime? lastDataDownload;
     final lastDateData = preferences.lastStationDataDate;
     log('last data ${lastDateData.toString()}');
     for (int i = 7; i >= 0; --i) {
@@ -196,30 +221,9 @@ class DataNotifier extends ChangeNotifier {
         final response = await http.get(url);
 
         if (response.statusCode == 200) {
-          final cvsResult = const CsvToListConverter().convert(
-            response.body,
-            fieldDelimiter: ';',
-            eol: '\n',
-            shouldParseNumbers: false,
-          );
-
-          if (cvsResult.length > 1) {
-            if (cvsResult[0].length > 2) {
-              cvsResult.removeAt(0);
-              final listData = cvsResult.map<DataStation>(
-                (line) {
-                  return DataStation.fromList(line);
-                },
-              ).toList();
-
-              listData.forEach((d) async {
-                await databaseHelper.insertStationData(d);
-              });
-
-              lastDataDownload = dateTime;
-              log('get data OK');
-            }
-          }
+          _decodeStationData(response.body);
+          lastDataDownload = dateTime;
+          log('get data OK');
         }
       }
     }
@@ -227,6 +231,12 @@ class DataNotifier extends ChangeNotifier {
     if (lastDataDownload != null) {
       preferences.setLastStationDataDate(lastDataDownload);
     }
+  }
+
+  Future<void> _readTestStationData() async {
+    final data =
+        await rootBundle.loadString('assets/test_data/nivo.202102.txt');
+    _decodeStationData(data);
   }
 
   Future<void> _initStation() async {
