@@ -7,8 +7,8 @@ import 'package:dart_rss/dart_rss.dart';
 class AvalancheAPI {
   final client = http.Client();
 
-  Future<AtomFeed> getAvalanche() async {
-    AtomFeed feed;
+  Future<AtomFeed?> getAvalanche() async {
+    AtomFeed? feed;
     try {
       final response =
           await client.get(Uri.parse('http://www.data-avalanche.org/feed'));
