@@ -3,12 +3,12 @@ import 'package:flutter/material.dart';
 import 'package:url_launcher/url_launcher.dart' as url;
 
 class MapLicenceWidget extends StatelessWidget {
-  const MapLicenceWidget({Key key}) : super(key: key);
+  const MapLicenceWidget({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    final bodyText2Color = Theme.of(context).textTheme.bodyText2.color;
-    final bodyText1Color = Theme.of(context).textTheme.bodyText1.color;
+    final bodyText2Color = Theme.of(context).textTheme.bodyText2?.color;
+    final bodyText1Color = Theme.of(context).textTheme.bodyText1?.color;
     return Container(
       color: Theme.of(context).backgroundColor,
       child: RichText(
