@@ -4,14 +4,17 @@ import 'package:snow_weather_info/modules/data_station/view.dart';
 import 'package:snow_weather_info/ui/nivose_page.dart';
 
 class StationCard extends StatelessWidget {
-  const StationCard({Key key, this.station}) : super(key: key);
+  const StationCard({
+    Key? key,
+    required this.station,
+  }) : super(key: key);
 
   final AbstractStation station;
 
   @override
   Widget build(BuildContext context) {
     var _textStyle = TextStyle(
-        color: Theme.of(context).textTheme.bodyText1.color,
+        color: Theme.of(context).textTheme.bodyText1?.color,
         fontStyle: FontStyle.normal);
     var snowHeigth = '';
     if (station is Station) {
@@ -20,7 +23,7 @@ class StationCard extends StatelessWidget {
           ? ' ${(st.lastSnowHeight * 100).toStringAsFixed(0)}cm'
           : '';
       final _color = st.hasData
-          ? Theme.of(context).textTheme.bodyText2.color
+          ? Theme.of(context).textTheme.bodyText2?.color
           : Theme.of(context).disabledColor;
       final _font = st.hasData ? FontStyle.normal : FontStyle.italic;
       _textStyle = TextStyle(color: _color, fontStyle: _font);
